Rename PurchasePremium model. Extend UserAdmin
Rename to PremiumPurchase as it makes more sense and reads better.
Previously created a UserAdmin class to customize admin view but turns
out was already a UserAdmin class so it overwrote it and caused all
manner of issues.
To fix, have instead extended UserAdmin with CustomUserAdmin and all is
well again. For now.