Fix bug with premium payment processing
Was possible for app to create a 'PremiumPurchase' object if a users
card failed.
Whilst this wouldn't grant them access to premium features, it meant
that if they tried again to pay, this time successfully, the app would
crash.
This is due to the 'user' field within the 'PremiumPurchase' model
having  a 1-1 relationship with the user model, causing a 'UNIQUE
constraint failed' exception.