Append _utils to utils scripts
Confusion creeping in so for all helper scripts add _utils to the end of
the name so as to clearly see it refers to the helper script and not
something else.
Link add collection button to view and it's own helpter function ready
to implement logic for a new collection.