Add 'set_page_name' helper function
Top nav bar contains an 'add bookmark' button which when linked
caused the settings pages to fail. This is because settings pages render
the top nav with the add bookmark button so they were looking for a
reference to page name, which wasn't there.
This function adds a check to pull page name from session, or if not
found to default to the users first page.