Add login_required dec's, refactor page check code
'login_required' decorators added to views that require a user be logged
in to function correctly.
Note, not yet added to all views that require it, working through entire
links.views.py file.
Updated page check code to redirect app to 'start_app' function in the
event of the requested page not being found. This is mostly to prevent
users accessing pages / areas they shouldn't via the address bar.