Finish updating views w/ direct access protection
All views within the links app should now be protected from users
trying to type in and access the view directly. In most cases it will
rediect them back to the main app page, and in some cases display a
message explaining why.
Also has the added benefit of recovering from errors by redirecting to
the main screen with an error message, rather than a 404.