Fix bug with saving resized images to AWS S3
When saving to AWS S3 using '.path', app was generating this error -
'NotImplementedError: This backend doesn't support absolute paths.'
This only happened with S3 and not locally. To fix, bypassed the need to
use .path when saving by instead saving a copy of the resized file, then
saving that copy.