Update main.js to fix error displaying invalid img
If the app fails to scrape an image from the url, usually because of a
faulty or incompatible image, it will return an emptry string and
display the missing file icon. Update main.js to check for this empty
string and instead display an image to tell user that an icon / image
could not be scraped.