Fix bug with favicon returning 404's
Favicon library will sometimes return urls to icons that resolve to 404.
When generating image from this url, it would just display a browser
default missing image icon. To get around this, before converting url to
a base64 string, requests is used to check for a 200 response to the
icon url before proceeding. If no 200, will return an empty string.