Update request headers when scraping urls
Found that some sites would return a 406 when trying to scrape (looking
at you, Uber) so to fix have updated headers with default HTTP Accept
header values for the chrome broswer.
Also, moved this into the .conf settings incase I need to update again,
as it's used in multiple places.