Update scraping to save image in correct format
Found bug whereby images are saved as png, regardless of format.
DIdn't seem to be an issue when displaying as browser would identify
correct format without using extension, but could not guarantee this
would work for all users.
Add 'image_ext' var to track the file format and this value is applied
when putting image together to ensure correct format is declared.