Fix issue with jQuery UI draggable on mobile
jQuery UI draggable isn't compatible with touch devices, it turns out.
The touch-punch library fixes this, by making the draggable elements
draggable on touch, but has the side effect of disabling clicks on said
elements.
To fix, have added touch punch, but also using .hande to specify the
draggable portion of the element.
Additionally, have also added a button for page sorting to turn this
handle on and off.