Update get_site_icon function to use regex
Found issue where searching for icons with the text 'icon' would also
return 'favicon' if that was found before 'icon'.
To fix, have changed the filename search to use regex so the list of
filenames can contain either strings, or regex statements.
In the case of searching for 'icon', now using negative look ahead to
remove 'favicon' from the search.