Update edit-bookmark error handling
On form errors, images now either reset to the icon stored in the db
for errors with the icon itself, or for other errors (title, url, etc)
the current image will be reloaded back into the form.
This avoids the user having to reselect images each time there is a
non-image related error.