Add link to main app sidebar icon
Originally added a link to the site icon in the sidebar, directing the
user to the main app page, for only the settings sidebar, and left out
main app as it seemed pointless at the time to have a link to the page
you are on.

Adding now as it will allow users to navigate back to the main app
from pages like add bookmark, arrange collections, and so on - which
isn't pointless.