Update app to adjust columns based on width
At certain breakpoints, if the number of columns are deemed as too
many for the current width, the user will be advised of this and
given the option of ignoring, or changing the number of columns.

The 'change-columns' button will automatically adjust the current
pages columns to the recommended number.