Fix bug with ise500 on deployed site w/ debug off
Bug only caused an error on deployed site with debug off, for reasons
still unknown.
Issue was with the template creating a url to the banner ad - url would
create fine locally, but would error on remote deployment if debugging
was off.
Fixed by using a template filter to create entire path to image, rather
than just generating image number and trying to slot it into template
with rest of the path.