Add tippy.js script to support & about pages
Short term fix. For reasons as yet unknown, when tippy scripts are not
loaded on a page, when clicking the search icon on Safari mobile,
nothing happens. If tippy is added, it works fine.
Can get around by adding 'cursor: pointer' and the click works, but then
clicking again will not remove the search input field. All very strange.
Using the 'tippy fix' for now until it can be investigated proper.